Stats refactored: use int instead of float

+++ Assets/_root/_src/Interfaces.cs
@@ -2,15 +2,15 @@ using UnityEngine;
 
 public interface IDamageDealer
 {
-    float DealDamage(IDamageTaker target, Damage dmg);
+    int DealDamage(IDamageTaker target, Damage dmg);
 }
 
 public interface IDamageTaker
 {
     Vector3 GetRootPosition();
     Vector3 GetHitPosition();
-    float TakeDamage(IDamageDealer origin, Damage dmg);
-    float TakeFatalDamage(IDamageDealer origin);
+    int TakeDamage(IDamageDealer origin, Damage dmg);
+    int TakeFatalDamage(IDamageDealer origin);
 }
 
 public interface IRaceAura
